[feat/#114]: 피드백 반영
